Make welcome screen fully overlay home screen
