+++ ui/i18n/src/main/res/values/strings.xml
@@ -59,7 +59,7 @@
     <string name="home_new_title">Review the new</string>
     <string name="home_downloads_title">Manage downloads</string>
     <string name="home_welcome_title">Welcome to AntennaPod!</string>
-    <string name="home_welcome_text">You are not subscribed to any podcasts yet.</string>
+    <string name="home_welcome_text">You are not subscribed to any podcasts yet. Press the menu button to add a podcast.</string>
     <string name="navigate_arrows">%s »</string>
     <string name="configure_home">Configure Home Screen</string>
